Jetpack Cooldown
Added a simple cooldown system to the jetpack (its not really a jetpack
btw, more of a pulsepack). It's indicated by a slider at the bottom
left. I wanted to make the slider change color when the pack can't be
used, but that doesn't seem to work the way I'm used to.