Updated end of level animation to teleport the block (with different options depending on which we prefer)
Also fixed a bug with loading scenes. The scenemanager.getactive()
cannot read folders in the build settings, so it thinks any level 2 is
the tutorial level 2 (since it's first in the settings). The only way
around this is to name our scenes differently or load by name. I chose
the latter